pan: Tweak the algorithm.
Null nodes and nodes which are nullable and contain parsed trees are
different. By carefully separating them and forcing trees during certain
steps, we can get Cat to actually carry information on its first language, so
that it yields useful trees.

Also a bunch of tiny fixes here and there.